orchestration-tools/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestration-tools"
version = "0.1.0"
description = "Verification and orchestration tools for the EmailIntelligence project"
readme = "README.md"
authors = [{name = "EmailIntelligence Team"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "GitPython>=3.1.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
    "cryptography>=42.0.0",
    "PyGithub>=1.59.0",
    "fastapi>=0.100.0",
    "pydantic>=2.11.5",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "mypy>=1.16.0",
    "pylint>=3.3.7",
]